backend/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
blinker==1.9.0
certifi==2026.2.25
click==8.1.8
coverage==7.10.7
distro==1.9.0
exceptiongroup==1.3.1
fastapi==0.104.1
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
groq==0.4.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.0
pluggy==1.6.0
pydantic==2.12.5
pydantic_core==2.41.5
pytest==7.4.3
pytest-cov==4.1.0
python-dotenv==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.48
starlette==0.27.0
tomli==2.4.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.24.0
Werkzeug==3.1.6
zipp==3.23.0
